python_exercise/requirements.txt
Django==2.1.7
djangorestframework==3.9.1
django-redis==4.10.0
pytz==2018.9
redis==3.2.0
requests==2.21.0
msgpack==0.6.1
django-environ==0.4.5